string-to-integer (stoi) changed to string-to-double (stod) for program input parameter thats supposed to parse doubles

+++ programs/cpp_version/craps/oop_version/main.cc
@@ -34,7 +34,7 @@ int main(int argc, char *argv[]) {
         numSamples = stoi(argv[1]);
         strategyInt = stoi(argv[2]);
         startingCash = atoi(argv[3]);
-        goalMultiplier = atoi(argv[4]);
+        goalMultiplier = stod(argv[4]);
 
     }
     catch (string E) {
